pyproject.toml
[tool.poetry]
name = "sdss-fliswarm"
version = "0.1.0-alpha.0"
description = "A tool to create and manage Docker instances of flicamera"
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/fliswarm"
repository = "https://github.com/sdss/fliswarm"
documentation = "https://sdss-fliswarm.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "fliswarm" }
]
include = ["fliswarm/etc/*", "LICENSE.md"]

[tool.poetry.scripts]
fliswarm = "fliswarm.__main__:main"

[tool.poetry.dependencies]
python = "^3.7"
sdsstools = ">=0.4.0"
click = "^7.1.2"
docker = "^4.3.1"
sdss-clu = "^0.5.2"
click-default-group = "^1.2.2"
daemonocle = "^1.0.2"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
matplotlib = ">=3.1.1"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = "^3.3.1"
Sphinx = "^3.0.0"
autopep8 = "^1.5.4"

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "fliswarm"
known_sdss = ["sdsstools"]
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
addopts = "--cov fliswarm --cov-report xml --cov-report html --cov-report term"

[tool.coverage.run]
branch = true
include = ["python/fliswarm/*"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=49.6.0"]
build-backend = "poetry.core.masonry.api"
